// frontend/sdk/models/MessageDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Unpaused Spotify API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MessageDto
 */
export interface MessageDto {
    /**
     * 
     * @type {string}
     * @memberof MessageDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof MessageDto
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof MessageDto
     */
    note: string;
    /**
     * 
     * @type {Date}
     * @memberof MessageDto
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof MessageDto
     */
    updatedAt: Date;
}

/**
 * Check if a given object implements the MessageDto interface.
 */
export function instanceOfMessageDto(value: object): value is MessageDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('note' in value) || value['note'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('updatedAt' in value) || value['updatedAt'] === undefined) return false;
    return true;
}

export function MessageDtoFromJSON(json: any): MessageDto {
    return MessageDtoFromJSONTyped(json, false);
}

export function MessageDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): MessageDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'title': json['title'],
        'note': json['note'],
        'createdAt': (new Date(json['createdAt'])),
        'updatedAt': (new Date(json['updatedAt'])),
    };
}

export function MessageDtoToJSON(json: any): MessageDto {
    return MessageDtoToJSONTyped(json, false);
}

export function MessageDtoToJSONTyped(value?: MessageDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'note': value['note'],
        'createdAt': value['createdAt'].toISOString(),
        'updatedAt': value['updatedAt'].toISOString(),
    };
}