// frontend/sdk/models/CreateMessageDto.ts
/* tslint:disable */
 
/**
 * Unpaused Spotify API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateMessageDto
 */
export interface CreateMessageDto {
    /**
     * 
     * @type {string}
     * @memberof CreateMessageDto
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof CreateMessageDto
     */
    note: string;
}

/**
 * Check if a given object implements the CreateMessageDto interface.
 */
export function instanceOfCreateMessageDto(value: object): value is CreateMessageDto {
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('note' in value) || value['note'] === undefined) return false;
    return true;
}

export function CreateMessageDtoFromJSON(json: any): CreateMessageDto {
    return CreateMessageDtoFromJSONTyped(json, false);
}

export function CreateMessageDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateMessageDto {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'],
        'note': json['note'],
    };
}

export function CreateMessageDtoToJSON(json: any): CreateMessageDto {
    return CreateMessageDtoToJSONTyped(json, false);
}

export function CreateMessageDtoToJSONTyped(value?: CreateMessageDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'note': value['note'],
    };
}