// frontend/sdk/models/UpdateMessageDto.ts
/* tslint:disable */
 
/**
 * Unpaused Spotify API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateMessageDto
 */
export interface UpdateMessageDto {
    /**
     * 
     * @type {string}
     * @memberof UpdateMessageDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateMessageDto
     */
    note?: string;
}

/**
 * Check if a given object implements the UpdateMessageDto interface.
 */
export function instanceOfUpdateMessageDto(value: object): value is UpdateMessageDto {
    return true;
}

export function UpdateMessageDtoFromJSON(json: any): UpdateMessageDto {
    return UpdateMessageDtoFromJSONTyped(json, false);
}

export function UpdateMessageDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateMessageDto {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'] == null ? undefined : json['title'],
        'note': json['note'] == null ? undefined : json['note'],
    };
}

export function UpdateMessageDtoToJSON(json: any): UpdateMessageDto {
    return UpdateMessageDtoToJSONTyped(json, false);
}

export function UpdateMessageDtoToJSONTyped(value?: UpdateMessageDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'note': value['note'],
    };
}